pyproject.toml
[project]
name = "news-summarizer"
version = "0.1.0"
description = "Tool to summarize latest Google News"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0.11",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
    "Markdown>=3.6",
    "openai>=1.0.0"
]
